creating a custom hook for fetching games 8.10
